Refactor editor mocks into reusable classes for tests
Introduced `MockEditor` and `MockEditorSettings` as reusable mock implementations to simplify and standardize unit tests. Updated `MarkerTestBase` and related tests to use these shared mocks, reducing duplicate code and improving maintainability.
